move the read methods to their own file, cleanup state after finishing the upload to not get stuck in a dirty state
